requirements.txt
torch
torchvision
tensorflow
numpy
matplotlib
scikit-learn
pillow
tqdm